database magic is happening
FINALLY got the db to save data, is able to read traders. has wants and gives. added a working form component to add more traders to list.

need to fix wants and gives so that new trader has those things when submitted to db. so far just the name goes in. need to set home page to find all traders by name and list them out as hrefs. each href will show a trader table on screen with that traders info. hopefully... have not yet tested reading trader data for the table yet.